requirements.txt
fastapi[standard]
pydantic-settings
chromadb
supabase
google-generativeai
python-dotenv